Extract webrick launcher to a separate file
